requirements.txt
flask==2.3.3
confluent-kafka==2.3.0
python-dotenv==1.0.0
